+++ rust/guest/k16-init/src/main.rs
@@ -8,7 +8,7 @@ use kraft_std::prelude::*;
 
 const PROMPT: &[u8] = b"INIT> ";
 const NEWLINE: &[u8] = b"\n";
-const HELP: &[u8] = b"HELP\nCLEAR\nECHO\nTICKS\nUNAME\nCAT\n";
+const HELP: &[u8] = b"HELP\nCLEAR\nECHO\nTICKS\nUNAME\nCAT\nALLOC\n";
 
 #[no_mangle]
 pub extern "C" fn main() -> ! {
@@ -69,6 +69,7 @@ fn dispatch_command(stdout: io::Fd, command: Command<'_>) {
         Command::Ticks => run_ticks(stdout),
         Command::Uname => run_uname(stdout),
         Command::Cat => run_cat(stdout),
+        Command::AllocTest => run_alloc_test(stdout),
         Command::Echo(bytes) => {
             must_write(stdout, bytes);
             must_write(stdout, NEWLINE);
@@ -96,6 +97,17 @@ fn run_uname(stdout: io::Fd) {
 
 fn run_cat(stdout: io::Fd) {
     match process::run("/bin/cat.kx") {
+        Ok(_) => {}
+        Err(process::Error::Syscall(status)) => {
+            must_write(stdout, b"ERR ");
+            must_write(stdout, run_error_name(status));
+            must_write(stdout, NEWLINE);
+        }
+    }
+}
+
+fn run_alloc_test(stdout: io::Fd) {
+    match process::run("/bin/alloc-test.kx") {
         Ok(_) => {}
         Err(process::Error::Syscall(status)) => {
             must_write(stdout, b"ERR ");